third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-block-import-non-self.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'">
<link id="selfImport"    rel="import" href="resources/hello.html">
<link id="nonSelfImport" rel="import" href="http://localhost:8080/htmlimports/resources/cors-basic.cgi">
</head>
<body>
<pre id="result"></pre>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

if (window.selfImport.import != null)
    result.innerHTML += "PASS\n";
else
    result.innerHTML += "FAIL: The import should be allowed.\n";

if (window.nonSelfImport.import == null)
    result.innerHTML += "PASS\n";
else
    result.innerHTML += "FAIL: The import should be blocked.\n";
</script>
</body>
</html>
