third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-block-import-non-self-expected.txt

CONSOLE ERROR: line 6: Refused to load the script 'http://localhost:8080/htmlimports/resources/cors-basic.cgi' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

PASS
PASS
